+++ wither/src/model.rs
@@ -1,10 +1,11 @@
 //! Model related code.
 
+use std::collections::HashMap;
+
 use async_trait::async_trait;
 use mongodb::bson::oid::ObjectId;
 use mongodb::bson::{doc, from_bson, to_bson};
 use mongodb::bson::{Bson, Document};
-
 use mongodb::options;
 use mongodb::results::DeleteResult;
 use mongodb::{Collection, Database};
@@ -13,7 +14,9 @@ use serde::{de::DeserializeOwned, Serialize};
 use crate::common::IndexModel;
 use crate::cursor::ModelCursor;
 use crate::error::{Result, WitherError};
-use std::collections::HashMap;
+
+const MONGO_ID_INDEX_NAME: &str = "_id_";
+const MONGO_DIFF_INDEX_BLACKLIST: [&str; 3] = ["v", "ns", "key"];
 
 /// This trait provides data modeling behaviors for interacting with MongoDB database collections.
 ///
@@ -195,7 +198,7 @@ where
         let updated_doc = coll
             .find_one_and_replace(filter, instance_doc, Some(opts))
             .await?
-            .ok_or_else(|| WitherError::ServerFailedToReturnUpdatedDoc)?;
+            .ok_or(WitherError::ServerFailedToReturnUpdatedDoc)?;
 
         // Update instance ID if needed.
         if id_needs_update {
@@ -222,7 +225,7 @@ where
     /// error.
     async fn update(self, db: &Database, filter: Option<Document>, update: Document, opts: Option<options::FindOneAndUpdateOptions>) -> Result<Self> {
         // Extract model's ID & use as filter for this operation.
-        let id = self.id().ok_or_else(|| WitherError::ModelIdRequiredForOperation)?;
+        let id = self.id().ok_or(WitherError::ModelIdRequiredForOperation)?;
 
         // Ensure we have a valid filter.
         let filter = match filter {
@@ -264,7 +267,7 @@ where
             .await?
             .map(Self::instance_from_document)
             .transpose()?
-            .ok_or_else(|| WitherError::ServerFailedToReturnUpdatedDoc)?)
+            .ok_or(WitherError::ServerFailedToReturnUpdatedDoc)?)
     }
 
     /// Delete this model instance by ID.
@@ -272,7 +275,7 @@ where
     /// Wraps the driver's `Collection.delete_one` method.
     async fn delete(&self, db: &Database) -> Result<DeleteResult> {
         // Return an error if the instance was never saved.
-        let id = self.id().ok_or_else(|| WitherError::ModelIdRequiredForOperation)?;
+        let id = self.id().ok_or(WitherError::ModelIdRequiredForOperation)?;
         Ok(Self::collection(db).delete_one(doc! {"_id": id}, None).await?)
     }
 
@@ -306,26 +309,38 @@ where
     /// any indexes defined on this model with the backend.
     ///
     /// This routine will destroy any indexes found on this model's collection which are not
-    /// defined in the response from `Self.indexes()`.
+    /// defined in this model's `indexes` method.
     async fn sync(db: &Database) -> Result<()> {
-        let coll = db.collection(Self::COLLECTION_NAME);
-        sync_model_indexes(db, &coll, Self::indexes()).await?;
+        let coll = Self::collection(db);
+        let current_indexes = get_current_indexes(&db, &coll).await?;
+        sync_model_indexes(db, &coll, Self::indexes(), current_indexes).await?;
         Ok(())
     }
 
-    /// Get current collection indexes if any
-    async fn get_current_indexes(db: &Database) -> HashMap<String, IndexModel> {
-        // Get the existing indexes, this result is an error when there are no indexes
-        let list_index_result = db.run_command(doc! {"listIndexes": Self::COLLECTION_NAME}, None).await;
-
-        create_current_indexes_map(list_index_result)
+    /// Get current collection indexes, if any.
+    async fn get_current_indexes(db: &Database) -> Result<HashMap<String, IndexModel>> {
+        let coll = Self::collection(db);
+        get_current_indexes(db, &coll).await
     }
 }
 
-const MONGO_ID_INDEX_NAME: &str = "_id_";
-const MONGO_DIFF_INDEX_BLACKLIST: [&str; 3] = ["v", "ns", "key"];
+/// Get current collection indexes, if any.
+async fn get_current_indexes(db: &Database, coll: &Collection) -> Result<HashMap<String, IndexModel>> {
+    let list_indexes = match db.run_command(doc! {"listIndexes": coll.name()}, None).await {
+        Ok(list_indexes) => list_indexes,
+        Err(err) => match err.kind.as_ref() {
+            // The DB & or collection does not yet exist. Move on.
+            mongodb::error::ErrorKind::CommandError(err) if err.code == 26 => doc! {},
+            _ => return Err(err.into()),
+        },
+    };
+    Ok(build_index_map(list_indexes))
+}
 
-// Reference: https://github.com/mongodb/specifications/blob/master/source/index-management.rst#index-name-generation
+/// Generate an index name from the keys of the given document, matching the behavior of the
+/// index management spec.
+///
+/// https://github.com/mongodb/specifications/blob/master/source/index-management.rst#index-name-generation
 fn generate_index_name_from_keys(keys: &Document) -> String {
     let mut key = keys.iter().fold(String::from(""), |mut acc, (key, value)| {
         acc.push_str(&format!("{}_{}_", key, value.as_i32().unwrap_or(0)));
@@ -336,216 +351,154 @@ fn generate_index_name_from_keys(keys: &Document) -> String {
     key
 }
 
-fn create_current_indexes_map(list_index_result: std::result::Result<Document, mongodb::error::Error>) -> HashMap<String, IndexModel> {
-    let mut current_indexes_map: HashMap<String, IndexModel> = HashMap::new();
+/// Build a mapping of index names to their index models.
+///
+/// NOTE: this algorithm is sub-optimal and does not account for every possible error which may
+/// arise during index processing. We would like to see the MongoDB team add the index management
+/// commands back into their client, however this will do the trick for now. The only real concern
+/// there is that the algorithm is not resilient to unexpected schema changes coming from the mongo
+/// server. These changes are unlikely, but we are just documenting this fact here for posterity.
+fn build_index_map(list_index: Document) -> HashMap<String, IndexModel> {
+    // Unpack the cursor.
+    let cursor = match list_index.get("cursor") {
+        Some(cursor) => cursor,
+        None => return Default::default(),
+    };
+    let doc = match cursor.as_document() {
+        Some(doc) => doc,
+        None => return Default::default(),
+    };
+    // https://docs.mongodb.com/manual/reference/limits/#Number-of-Indexes-per-Collection
+    // We have a maximum of 64 indexes per collection, the firstBatch contains them all based on our tests.
+    let first_batch = match doc.get_array("firstBatch").ok() {
+        Some(first_batch) => first_batch,
+        None => return Default::default(),
+    };
 
-    if let Ok(doc) = list_index_result {
-        // There are indexes
-        if let Some(cursor) = doc.get("cursor") {
-            let doc = cursor.as_document().unwrap();
+    let index_map = first_batch
+        .iter()
+        // Extract documents.
+        .filter_map(|bson| bson.as_document().cloned())
+        // Filter out default index.
+        .filter(|doc| {
+            match doc.get_str("name").ok() {
+                Some(name) if name == MONGO_ID_INDEX_NAME => false, // Filter out.
+                Some(_) => true, // Include.
+                None => false, // Filter out.
+            }
+        })
+        .fold(HashMap::new(), |mut acc, doc| {
+            // Extract document keys & generate index name based on keys.
+            let idx_keys = match doc.get_document("key").ok() {
+                Some(idx_keys) => idx_keys,
+                None => return acc,
+            };
+            let index_name = generate_index_name_from_keys(idx_keys);
 
-            // https://docs.mongodb.com/manual/reference/limits/#Number-of-Indexes-per-Collection
-            // We have a maximum of 64 indexes per collection, the firstBatch contains them all in my tests
-            // If anyone is capable to understand how to make this batch overflow let me know and I'll fix it
-            let first_batch = doc.get_array("firstBatch").unwrap();
+            // Build index model, filtering out blacklisted keys.
+            let mut options = Document::new();
+            doc.iter().for_each(|(b_key, b_value)| {
+                if !MONGO_DIFF_INDEX_BLACKLIST.contains(&b_key.as_str()) {
+                    options.insert(b_key.to_string(), b_value);
+                }
+            });
+            let model = IndexModel::new(idx_keys.clone(), Some(options));
 
-            first_batch
-                .iter()
-                // Convert to document
-                .map(|bson| bson.as_document().unwrap().clone())
-                // Remove default index
-                .filter(|doc| doc.get_str("name").unwrap() != MONGO_ID_INDEX_NAME)
-                .for_each(|doc| {
-                    // Index keys
-                    let idx_keys = doc.get_document("key").unwrap();
-
-                    // Unique generated key for the map
-                    let key = generate_index_name_from_keys(idx_keys);
-
-                    // These are the options
-                    let mut options = Document::new();
-
-                    doc.iter().for_each(|(b_key, b_value)| {
-                        // Remove unused stuff (key, namespace, version)
-                        let key_in_blacklist = !MONGO_DIFF_INDEX_BLACKLIST
-                            .iter()
-                            .all(|blacklisted| b_key != blacklisted);
-                        if !key_in_blacklist {
-                            options.insert(b_key, b_value);
-                        }
-                    });
-
-                    // Custom IndexModel
-                    let model = IndexModel::new(idx_keys.clone(), Some(options));
-
-                    // Add the IndexModel to the current_indexes_map
-                    current_indexes_map.insert(key, model);
-                });
-        }
-    }
-
-    current_indexes_map
+            acc.insert(index_name, model);
+            acc
+        });
+    index_map
 }
 
-async fn sync_model_indexes<'a>(db: &'a Database, coll: &'a Collection, model_indexes: Vec<IndexModel>) -> Result<()> {
+async fn sync_model_indexes<'a>(
+    db: &'a Database, coll: &'a Collection, model_indexes: Vec<IndexModel>, current_indexes_map: HashMap<String, IndexModel>,
+) -> Result<()> {
     log::info!("Synchronizing indexes for '{}'.", coll.namespace());
 
-    // Get the existing indexes, this result is an error when there are no indexes
-    let list_index_result = db.run_command(doc! {"listIndexes": coll.name()}, None).await;
-
-    // The already present indexes
-    let current_indexes_map = create_current_indexes_map(list_index_result);
-
-    // Fetch target indexes for this model.
-    let mut target_indexes_map: HashMap<String, IndexModel> = HashMap::new();
-    for model in model_indexes.iter() {
+    // Build a mapping of aspired indexes based on the model's declared indexes.
+    let aspired_indexes_map = model_indexes.iter().fold(HashMap::new(), |mut acc, model| {
         let mut target_model = model.clone();
         // Populate the 'target' indexes map for easy comparison later.
         let key = generate_index_name_from_keys(&model.keys);
 
-        // If we have options
-        if let Some(ref mut options) = target_model.options {
-            // Try to get the name
-            if options.get_str("name").is_err() {
-                // If the name option is missing add it
-                options.insert("name", key.clone());
+        // Ensure we have an options object with at least the index name.
+        match &mut target_model.options {
+            Some(options) => {
+                if options.get_str("name").ok().is_none() {
+                    options.insert("name", key.clone());
+                }
             }
-        } else {
-            // If the model has no option then add the name
-            let options = doc! {"name": key.clone()};
-            target_model.options = Some(options);
+            // If no options are present, then add a default options doc with the index name.
+            None => {
+                let options = doc! { "name": key.clone() };
+                target_model.options = Some(options);
+            }
         }
-
-        target_indexes_map.insert(key, target_model.clone());
-    }
-
-    // These are the indexes that must be dropped (I just need the name)
-    let mut indexes_to_drop: Vec<String> = Vec::new();
-
-    // These are the indexes that must be created
-    let mut indexes_to_create: HashMap<String, IndexModel> = HashMap::new();
-
-    // First drop the additional current_indexes
-    current_indexes_map.iter().for_each(|(key, current_index)| {
-        // There is no key in target
-        if target_indexes_map.get(key).is_none() {
-            indexes_to_drop.push(String::from(current_index.options.as_ref().unwrap().get_str("name").unwrap()));
-        }
+        acc.insert(key, target_model);
+        acc
     });
 
-    // Iter over target_indexes
-    for (key, index_model) in target_indexes_map.iter() {
-        // Check if key already exists
-        if let Some(current_index) = current_indexes_map.get(key) {
-            /*
-            // First check differences in the key
-            let current_index_keys = &current_index.keys;
-            let target_index_keys = &index_model.keys;
+    // For any current index which does not exist in the model's aspired indexes
+    // list, add it to the drop list.
+    let mut indexes_to_drop = current_indexes_map.iter().fold(vec![], |mut acc, (key, _)| {
+        if !aspired_indexes_map.contains_key(key) {
+            acc.push(key);
+        }
+        acc
+    });
 
-            // If it exist we must do some checks
-            let has_different_key =
-            // There is a difference if any current keys is different from any target keys
-            current_index_keys.iter().any(|(current_key, current_value)| {
-                if let Some(target_value) = target_index_keys.get(current_key) {
-                    // This is the order of the key (+1 or -1)
-                    target_value != current_value
-                } else {
-                    // Added a key to the index keys
-                    false
-                }
-            });
-            // There is a difference, drop the index
-            if has_different_key {
-                indexes_to_drop.push(String::from(
-                    current_index
-                        .options
-                        .as_ref()
-                        .unwrap()
-                        .get_str("name")
-                        .unwrap(),
-                ));
-                indexes_to_create.insert(key.clone(), index_model.clone());
-            } else {
-            */
-            // We need to check options
-            // If the index is already present there must be the "name" option so "options" is always defined
-            let current_index_options_doc = &current_index.options.as_ref().unwrap();
-            // The target index could have no options
-            let target_index_options_doc = &index_model.options.as_ref().unwrap();
-
-            // if let Some(target_index_options_doc) = target_index_options {
-            // Iter over target_index_options_doc
-            let has_diff = target_index_options_doc.iter().any(|(target_key, target_value)| {
-                // If the target option is also in current index option
-                let current_index_option = current_index_options_doc.get(target_key);
-                // We have a diff if ANY current option is not equal to the target option
-                if let Some(current_value) = current_index_option {
-                    current_value != target_value
-                } else {
-                    true
-                }
-            });
-            if has_diff {
-                indexes_to_drop.push(String::from(current_index_options_doc.get_str("name").unwrap()));
-                indexes_to_create.insert(key.clone(), index_model.clone());
+    // Diff aspired indexes with current indexes, and update our lists of indexes to create and
+    // drop based on diffing the options of each index model. This is based purely on the
+    // implementation of PartialEq on the bson::Document type.
+    let mut indexes_to_create: HashMap<String, IndexModel> = HashMap::new();
+    for (aspired_index_name, aspired_index) in aspired_indexes_map.iter() {
+        // Unpack the corresponding current index by name if it exists, else prep it for creation.
+        let current_index = match current_indexes_map.get(aspired_index_name) {
+            Some(current_index) => current_index,
+            // If the aspired index does not exist by name on the collection,
+            // then we need to create it.
+            None => {
+                indexes_to_create.insert(aspired_index_name.clone(), aspired_index.clone());
+                continue;
             }
-        // } else {
-        //     // This can be tricky, if the target_index has no options it could also mean that the user did not specify a name
-        //     // But in the current_index there is always at least the name which is generated from the key
-        //     // let has_custom_options = current_index_options_doc.iter().any(|(k, _)| {
-        //     //     k != "name"
-        //     // });
-        //     // if has_custom_options {
-        //     indexes_to_drop.push(String::from(
-        //         current_index_options_doc.get_str("name").unwrap(),
-        //     ));
-        //     indexes_to_create.insert(key.clone(), index_model.clone());
-        //     // }
-        // }
-        // }
-        } else {
-            // If it doesn't just add it
-            indexes_to_create.insert(key.clone(), index_model.clone());
+        };
+
+        // If the options of the two index models do not match, then we need to drop the existing
+        // and create an updated version.
+        if aspired_index.options != current_index.options {
+            indexes_to_drop.push(aspired_index_name);
+            indexes_to_create.insert(aspired_index_name.clone(), aspired_index.clone());
         }
     }
 
-    // `To drop multiple indexes (Available starting in MongoDB 4.2), specify an array of the index
-    // names` If we drop support for MongoDB <= 4.2 we could drop all indexes at once
-    for drop_key in indexes_to_drop {
+    // Drop indexes which have been flagged for dropping.
+    for index_name in indexes_to_drop {
         let drop_command = doc! {
             "dropIndexes": coll.name(),
-            "index": drop_key
+            "index": index_name,
         };
-
         db.run_command(drop_command, None).await?;
     }
 
-    let mut create_command = doc! {
-        "createIndexes": coll.name(),
-    };
-
-    let mut indexes_doc: Vec<Document> = Vec::new();
-
-    // Foreach index prepare the documents
-    for (_, model) in indexes_to_create.iter() {
-        let mut options = Document::new();
-        if let Some(opts) = model.options.clone() {
-            options = opts;
+    // Create any indexes which have been flagged for creation.
+    let indexes_to_create = indexes_to_create.into_iter().fold(vec![], |mut acc, (_, index_model)| {
+        let mut index_doc = Document::new();
+        index_doc.insert("key", index_model.keys);
+        if let Some(options) = index_model.options {
+            index_doc.extend(options);
         }
-
-        let key = model.keys.clone();
-
-        let mut doc_index = Document::new();
-
-        doc_index.insert("key", key);
-        doc_index.extend(options);
-
-        indexes_doc.push(doc_index);
-    }
-
-    if !indexes_doc.is_empty() {
-        create_command.insert("indexes", indexes_doc);
-        db.run_command(create_command, None).await?;
+        acc.push(index_doc);
+        acc
+    });
+    if !indexes_to_create.is_empty() {
+        db.run_command(
+            doc! {
+                "createIndexes": coll.name(),
+                "indexes": indexes_to_create,
+            },
+            None,
+        )
+        .await?;
     }
 
     log::info!("Synchronized indexes for '{}'.", coll.namespace());